State Machines, Fixed Components
A little delve into XState machines, tried implementing them with a scroll tracking service, but realised it was slightly too complex for such a trivial problem.

Instead resulted to using vuex for a basic store to track scrolling events, homepage now has a finished component for contacting me, which appears after the user has stopped scrolling for a period of time.