Page Transitions
After many, many hours, a pizza and 2 coffees later I have finally settled on a basic fluid page transition.

Using a lottie animation, which is new for me. It isn't 100%, animations are a clunky area for me. Ended up adding delays to routing to account for page changes mid animation, messy but works.

Now I have to actually get on with page content...